last changes

+++ controllers/item.js
@@ -99,39 +99,48 @@ const searchItems = async (req, res) => {
 
 const updateItem = async (req, res) => {
     try {
-        const { name, picture, price, description, pieceCount, theme } = req.body;
-        const updatedItem = await itemService.updateItem(req.params.id, {
+        const { id } = req.params;
+        const { name, description, price, picture, theme, pieces } = req.body;
+        
+        const updatedItem = await itemService.updateItem(id, {
             name,
-            picture,
-            price: Number(price),
             description,
-            pieceCount: Number(pieceCount),
-            theme
+            price: Number(price),
+            picture,
+            theme,
+            pieces: Number(pieces)
         });
+
         if (!updatedItem) {
-            return res.status(404).render('error', { message: 'Item not found' });
+            return res.status(404).json({ message: 'Item not found' });
         }
-        res.redirect('/');
+
+        res.status(200).json({ message: 'Item updated successfully', item: updatedItem });
     } catch (error) {
-        res.status(400).render('error', { message: 'Error updating item', error: error.message });
+        res.status(400).json({ message: 'Error updating item', error: error.message });
     }
 };
 
+
 const deleteItem = async (req, res) => {
     try {
-        const deletedItem = await itemService.deleteItem(req.params.id);
+        const { id } = req.params;
+        if (!req.session.isAdmin) {
+            return res.status(403).json({ message: 'You must be an admin to delete items' });
+        }
+        const deletedItem = await itemService.deleteItem(id);
         if (!deletedItem) {
-            return res.status(404).render('error', { message: 'Item not found' });
+            return res.status(404).json({ message: 'Item not found' });
         }
-
+        
         const cartsResult = await cartService.deleteItemFromAllCarts(deletedItem._id);
-        if (!cartsResult){
-            return res.status(404).render('error', {message: 'Error Deleting Items from all carts.'})
+        if (!cartsResult) {
+            return res.status(500).json({ message: 'Error deleting items from all carts' });
         }
 
-        res.redirect('/');
+        res.status(200).json({ message: 'Item deleted successfully' });
     } catch (error) {
-        res.status(500).render('error', { message: 'Error deleting item', error: error.message });
+        res.status(500).json({ message: 'Error deleting item', error: error.message });
     }
 };
